Semantics and Styling Change
Converted HTML id's to kebab case, and classes to follow BEM syntax. All scripts and specs were refactored to account for this. Created a cookies-per-second element that tracks how many cookies are generated by clicking companions per second (does not factor in the actual users clicks). Made the body fade in on load

Signed-off-by: Colin Wakefield <59844720+ColinWake@users.noreply.github.com>